Move script crate unit tests into the unit_tests crate.
