+++ tests/unit/lib.rs
@@ -9,6 +9,7 @@ extern crate gfx;
 extern crate net;
 extern crate net_traits;
 extern crate profile;
+extern crate script;
 extern crate selectors;
 extern crate string_cache;
 extern crate style;
@@ -17,5 +18,6 @@ extern crate url;
 
 #[cfg(test)] #[path="gfx/mod.rs"] mod gfx_tests;
 #[cfg(test)] #[path="net/mod.rs"] mod net_tests;
+#[cfg(test)] #[path="script/mod.rs"] mod script_tests;
 #[cfg(test)] #[path="style/mod.rs"] mod style_tests;
 #[cfg(test)] #[path="util/mod.rs"] mod util_tests;
