+++ tests/unit/script/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(target_pointer_width = "64")] mod size_of;
+mod textinput;
